requirements.txt
Flask
langchain
openai
chromadb
beautifulsoup4
requests
python-dotenv
faiss-cpu
transformers
sentence-transformers
webdriver-manager
selenium
langchainhub 
google-api-python-client 
huggingface_hub
langchainhub
langchain_google_community
groq